telegram_bot/requirements.txt
# Build tools (cần cài trước)
setuptools>=65.0.0
wheel>=0.38.0

# Telegram Bot dependencies
python-telegram-bot>=21.0
flask>=3.0.0
python-dotenv>=1.0.0

